moved colored logging config to utils
