+++ gunicorn_logging.conf
@@ -45,6 +45,6 @@ formatter=generic
 args=('./logs/gunicorn.access.log', 'a', 10485760)
 
 [formatter_generic]
-format=%(asctime)s [%(process)d:%(name)s:%(lineno)s] [%(levelname)s] %(message)s
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s (%(filename)s:%(lineno)d)
 datefmt=%Y-%m-%d %H:%M:%S
-class=colored_logging.ColorFormatter
+class=apis.utils.ColorFormatter
